Moves function to utils
